# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /videos/{videoId}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get video details with lock status
      parameters:
        - in: path
          name: videoId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Video details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoDetailResponse"
        "403":
          description: Not enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /progress/videos/{videoId}:
    get:
      operationId: getVideoProgress
      tags: [progress]
      summary: Get video progress
      parameters:
        - in: path
          name: videoId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Video progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoProgressResponse"

    post:
      operationId: updateVideoProgress
      tags: [progress]
      summary: Update video progress
      parameters:
        - in: path
          name: videoId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgressRequest"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoProgressResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    VideoDetailResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        youtubeUrl:
          type: string
        orderIndex:
          type: integer
        durationSeconds:
          type: integer
          nullable: true
        subjectId:
          type: integer
        sectionId:
          type: integer
        sectionTitle:
          type: string
        previousVideoId:
          type: integer
          nullable: true
        nextVideoId:
          type: integer
          nullable: true
        locked:
          type: boolean
        unlockReason:
          type: string
          nullable: true
      required:
        - id
        - title
        - youtubeUrl
        - orderIndex
        - subjectId
        - sectionId
        - sectionTitle
        - locked
    VideoProgressResponse:
      type: object
      properties:
        videoId:
          type: integer
        lastPositionSeconds:
          type: integer
        isCompleted:
          type: boolean
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - videoId
        - lastPositionSeconds
        - isCompleted
    UpdateProgressRequest:
      type: object
      properties:
        lastPositionSeconds:
          type: integer
        isCompleted:
          type: boolean
      required:
        - lastPositionSeconds
...